changed version from long to int

+++ webofneeds/won-core/src/main/java/won/protocol/model/Connection.java
@@ -38,7 +38,7 @@ public class Connection implements ParentAware<ConnectionContainer>
 
   @Version
   @Column(name="version", columnDefinition = "integer DEFAULT 0", nullable = false)
-  private long version = 0L;
+  private int version = 0;
 
   @Version
   @Temporal(TemporalType.TIMESTAMP)
@@ -183,11 +183,11 @@ public class Connection implements ParentAware<ConnectionContainer>
     this.state = state;
   }
 
-  protected void setVersion(final long version) {
+  protected void setVersion(final int version) {
     this.version = version;
   }
 
-  public long getVersion() {
+  public int getVersion() {
     return version;
   }
